Fix Mac build error due to -Wsign-compare. (#2979)

+++ sky/shell/systrace_logger.cc
@@ -13,7 +13,7 @@
 namespace sky {
 namespace shell {
 
-static const size_t kBufferSize = 256;
+static const int kBufferSize = 256;
 
 SystraceLogger::SystraceLogger()
     : trace_fd_(HANDLE_EINTR(
